提交任务倒计时

+++ KaiWan/TaskInfoView.m
@@ -84,7 +84,7 @@
         for (int i = 0; i < rewardArr.count; i++) {
             sumReward += [rewardArr[i] integerValue];
         }
-        self.moneyLabel.text = [NSString stringWithFormat:@"+%ld元", sumReward];
+        self.moneyLabel.text = [NSString stringWithFormat:@"+%ld元", (unsigned long)sumReward];
     } else {
         self.moneyLabel.text = [NSString stringWithFormat:@"+%@元", dataDic[@"reward"]];
     }
@@ -127,7 +127,7 @@
     self.stepView.frame = CGRectMake(0, CGRectGetMaxY(self.infoView.frame), SWIDTH, self.stepLabel.frame.size.height + HeightScale(30));
     
     self.frame = CGRectMake(0, 0, SWIDTH, self.stepView.frame.size.height + self.infoView.frame.size.height);
-  
+    
 }
